% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_auth}
\alias{asapi_auth}
\title{The Authentication Function}
\usage{
asapi_auth(email = NULL, password = NULL, url = "https://api.agrospace.cl")
}
\arguments{
\item{email}{Email of custome user}

\item{password}{Password of register user in https://api.agrospace.cl}

\item{url}{URL for dev purpose}
}
\description{
This function retrieve your AgroSpace API KEY
}
\examples{
asapi_auth(email="user.example@agrospace.cl", password="contra1234")

}
\keyword{API_KEY,}
\keyword{api_key}
