% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_client_get}
\alias{asapi_client_get}
\title{The Client GET Function}
\usage{
asapi_client_get(client, email, api_key, url = "https://api.agrospace.cl")
}
\arguments{
\item{client}{Client name}

\item{email}{email of user}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}
}
\description{
This function allows you to GET Client information with the API and retrieve your API KEY
}
\examples{
asapi_client_get(client="clientexample", email="user.example@agrospace.cl",
api_key=asapi_auth(email="user.example@agrospace.cl", password="contra1234")$api_key)

}
\keyword{api_key}
