% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_client_post}
\alias{asapi_client_post}
\title{The Client POST Function}
\usage{
asapi_client_post(
  client_name,
  email,
  contact,
  api_key,
  url = "https://api.agrospace.cl"
)
}
\arguments{
\item{client_name}{Name of de new client.}

\item{email}{Email of the contact of the new client.}

\item{contact}{Contact name.}

\item{api_key}{AgroSpace API Key.}

\item{url}{URL for dev purpose}
}
\description{
This function allows you to POST New Client with the API and retrieve your API KEY
}
