% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_construct}
\alias{asapi_construct}
\title{The API Constructor Function}
\usage{
asapi_construct(param_query = NULL, api_key = NULL)
}
\arguments{
\item{param_query}{List of key parameters.}

\item{api_key}{API key return by asapi_auth() function.}
}
\description{
Generical function to add api key token for each call
}
\examples{
asapi_construct()
}
\keyword{constructor}
