% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_farm_delete}
\alias{asapi_farm_delete}
\title{The Farms DELETE Function}
\usage{
asapi_farm_delete(
  client,
  farm,
  email,
  api_key,
  url = "https://api.agrospace.cl"
)
}
\arguments{
\item{client}{Client name}

\item{farm}{Farm name to query}

\item{email}{email of user}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}
}
\description{
This function allows you to DELETE Farms information with the API and retrieve your API KEY
}
\examples{
asapi_farm_delete(client="clientexample", farm="farmexample",
email="user.example@agrospace.cl", api_key=asapi_auth(email="user.example@agrospace.cl", password="contra1234")$api_key)
}
\keyword{Farms}
