% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_farm_post}
\alias{asapi_farm_post}
\title{The Farms POST Function}
\usage{
asapi_farm_post(
  client,
  farm_name,
  geojson = "NULL",
  email,
  api_key,
  url = "https://api.agrospace.cl"
)
}
\arguments{
\item{client}{Client name}

\item{farm_name}{Farm name to be set}

\item{geojson}{geojson apram}

\item{email}{email of user}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}
}
\description{
This function allows you to POST new Farm with the API and retrieve your API KEY
}
\keyword{Farms}
