% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_farm_put}
\alias{asapi_farm_put}
\title{The Farms PUT Function}
\usage{
asapi_farm_put(
  client,
  farm,
  new_type = "NULL",
  new_region = "NULL",
  new_commune = "NULL",
  new_geojson = "NULL",
  email_admin,
  api_key,
  url = "https://api.agrospace.cl"
)
}
\arguments{
\item{client}{Client name}

\item{farm}{Farm name to query}

\item{new_type}{New farm type. If the value is 'NULL' it does not change.}

\item{new_region}{New farm region. If the value is 'NULL' it does not change.}

\item{new_commune}{New farm Commune. If the value is 'NULL' it does not change.}

\item{new_geojson}{New farm geojson. If the value is 'NULL' it does not change.}

\item{email_admin}{Email of the user who is editing. It must be 'admin'.}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}
}
\description{
This function allows you to PUT farm new information with the API and retrieve your API KEY
}
\examples{
asapi_farm_put(client="clientexample", farm="newfarmexample",
new_region="Region de los Lagos", email_admin="user.example@agrospace.cl", api_key=asapi_auth(email="user.example@agrospace.cl", password="contra1234")$api_key)

}
\keyword{Farms}
