% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_features_get}
\alias{asapi_features_get}
\title{The Features GET Function}
\usage{
asapi_features_get(index, farm_type, api_key, url = "http://api.agrospace.cl")
}
\arguments{
\item{index}{index}

\item{farm_type}{farm type}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}
}
\description{
This function allows you to GET Features information like Index, Labels, Class, Pallete, others.
Agrospace members only, with apikey master.
}
\examples{
asapi_features_get(index="BIOMASS", farm_type="esmeralda", api_key="APIKEYEXAMPLE")
}
\keyword{API_KEY,}
\keyword{api_key}
