% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_features_post}
\alias{asapi_features_post}
\title{The Features POST Function}
\usage{
asapi_features_post(api_key, url = "https://api.agrospace.cl")
}
\arguments{
\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}
}
\description{
This function Update feature info from googledrive to mongoDB.
Agrospace members only, with apikey master.
}
