% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_image}
\alias{asapi_image}
\title{The image GET Function}
\usage{
asapi_image(
  client,
  farm,
  sensor,
  index,
  date,
  email,
  api_key,
  url = "http://api.agrospace.cl"
)
}
\arguments{
\item{client}{Client name}

\item{farm}{Farm name to query}

\item{sensor}{Sensor}

\item{index}{Index}

\item{date}{imagen date}

\item{email}{email of user}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}
}
\description{
This function allows you to GET Client - Farm raster information with the API
}
\examples{
asapi_image()
}
\keyword{API_KEY,}
\keyword{api_key}
