% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_index_get}
\alias{asapi_index_get}
\title{The Index GET Function}
\usage{
asapi_index_get(
  client,
  email,
  farm,
  sensor,
  api_key,
  url = "https://api.agrospace.cl",
  dash_param = FALSE
)
}
\arguments{
\item{client}{Client name}

\item{email}{email of}

\item{farm}{farm name}

\item{sensor}{sensor}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}

\item{dash_param}{AgroSpace internal use parameter. Default value FALSE}
}
\description{
This function allows you to GET Index information of client
}
