% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_json}
\alias{asapi_json}
\title{JSON editor form}
\usage{
asapi_json(res = NULL)
}
\arguments{
\item{res}{Respone from api.agrospace.cl}
}
\description{
Generical function to translate json to data.frame
}
