% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_layout}
\alias{asapi_layout}
\title{The Layout image GET Function}
\usage{
asapi_layout(client, farm, url = "https://api.agrospace.cl")
}
\arguments{
\item{client}{Client name}

\item{farm}{Client name}

\item{url}{URL for dev purpose}
}
\description{
This function returns the url of a farm's layout with the API
}
\examples{
asapi_layout(client="clientexample", farm="farm1example")
}
\keyword{API_KEY,}
\keyword{api_key}
