% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_layout_post}
\alias{asapi_layout_post}
\title{The Layout image POST Function}
\usage{
asapi_layout_post(
  client,
  path,
  email,
  api_key,
  url = "https://api.agrospace.cl"
)
}
\arguments{
\item{client}{Client name.}

\item{path}{Address of the new logo.}

\item{email}{email of user}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}

\item{farm}{Farm name.}
}
\description{
This function upload a farms's layout with the API
}
\examples{
asapi_layout_post(client="clientexample", farm="farm1example", path="https://pbs.twimg.com/profile_images/1063094754228011008/-AN6PKnB.jpg", email="user.example@agrospace.cl", api_key="APIKEYEXAMPLE")
}
\keyword{API_KEY,}
\keyword{api_key}
