% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_logo}
\alias{asapi_logo}
\title{The Logo image GET Function}
\usage{
asapi_logo(client, url = "https://api.agrospace.cl")
}
\arguments{
\item{client}{Client name}

\item{url}{URL for dev purpose}
}
\description{
This function returns the url of a client's logo with the API
}
