% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_logo_post}
\alias{asapi_logo_post}
\title{The Logo image POST Function}
\usage{
asapi_logo_post(client, path, email, api_key, url = "https://api.agrospace.cl")
}
\arguments{
\item{client}{Client name.}

\item{path}{Address of the new logo.}

\item{email}{email of user}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}
}
\description{
This function upload a client's logo with the API
}
\examples{
asapi_logo_post(client="clientexample", path="https://pbs.twimg.com/profile_images/1063094754228011008/-AN6PKnB.jpg", email="user.example@agrospace.cl", api_key=asapi_auth(email="user.example@agrospace.cl", password="contra1234")$api_key)
}
\keyword{API_KEY,}
\keyword{api_key}
