% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_new_apikey}
\alias{asapi_new_apikey}
\title{The New ApiKey Function}
\usage{
asapi_new_apikey(
  email = NULL,
  password = NULL,
  url = "https://api.agrospace.cl"
)
}
\arguments{
\item{email}{Email of custome user}

\item{password}{Password of register user in https://api.agrospace.cl}

\item{url}{URL for dev purpose}
}
\description{
This function Reset your AgroSpace API Key, entering your session.
}
\examples{
asapi_new_apikey(email="user.example@agrospace.cl", password="contra1234")

}
\keyword{API_KEY,}
\keyword{api_key}
