% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_new_pass}
\alias{asapi_new_pass}
\title{The New Password Function}
\usage{
asapi_new_pass(
  email = NULL,
  old_password = NULL,
  new_password = NULL,
  url = "https://api.agrospace.cl"
)
}
\arguments{
\item{email}{Email of custome user}

\item{old_password}{Password of register user in https://api.agrospace.cl}

\item{new_password}{New Password user}

\item{url}{URL for dev purpose}
}
\description{
This function Reset your user Password, entering your session.
}
\keyword{API_KEY,}
\keyword{api_key}
