% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_planes_get}
\alias{asapi_planes_get}
\title{The plan GET Function}
\usage{
asapi_planes_get(plan, api_key, url = "https://api.agrospace.cl")
}
\arguments{
\item{plan}{Name of the plan. You can use 'ALL' to call all available plans.}

\item{api_key}{AgroSpace API Key.}

\item{url}{URL for dev purpose}
}
\description{
This function returns information on plans, with the API.
}
\keyword{API_KEY,}
\keyword{api_key}
