% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_table_cover}
\alias{asapi_table_cover}
\title{The Table Cover GET Function}
\usage{
asapi_table_cover(
  client,
  farm,
  tableid,
  sensor,
  index,
  date_start,
  date_end,
  email,
  api_key,
  url = "http://api.agrospace.cl"
)
}
\arguments{
\item{client}{Client name}

\item{farm}{Farm name to query}

\item{tableid}{The mean values for the field (levelzero) or for each paddock (levelone)}

\item{sensor}{Sensor}

\item{index}{Index}

\item{date_start}{date from, YYYY-MM-DD format}

\item{date_end}{date to, YYYY-MM-DD format}

\item{email}{email of user}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}
}
\description{
This function allows you to GET Cover information with the API
}
\examples{
asapi_table_cover(client="clientexample", farm="farm1example", tableid="levelzero", sensor="S2SR", index="NDVI", date_start="2020-12-04", date_end="2020-12-19", email="user.example@agrospace.cl", api_key="APIKEYEXAMPLE", url="https://api.agrospace.cl")
}
\keyword{API_KEY,}
\keyword{api_key}
