% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_table_stats}
\alias{asapi_table_stats}
\title{The Table GET Function}
\usage{
asapi_table_stats(
  client,
  farm,
  datasetid,
  tableid,
  month,
  email,
  api_key,
  url = "https://api.agrospace.cl",
  dash_param = FALSE
)
}
\arguments{
\item{client}{Client name}

\item{farm}{Farm name to query}

\item{datasetid}{dataset prod or dev}

\item{tableid}{The mean values for the field (levelzero) or for each paddock (levelone)}

\item{month}{Boolean vaule TRUE or FALSE to retrieve monthly stats (TRUE)}

\item{email}{email of user}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}

\item{dash_param}{AgroSpace internal use parameter. Default value FALSE}
}
\description{
This function allows you to GET Client Farms  information with the API
}
