% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_terramodels_post}
\alias{asapi_terramodels_post}
\title{The Terramodels POST Function}
\usage{
asapi_terramodels_post(
  client,
  farm,
  sensor,
  index,
  date_start,
  date_end,
  type_export,
  email,
  api_key,
  dash_param = FALSE,
  url = "https://api.agrospace.cl"
)
}
\arguments{
\item{client}{Client name}

\item{farm}{Farm name to query}

\item{sensor}{Sensor}

\item{index}{Index}

\item{date_start}{date from, YYYY-MM-DD format}

\item{date_end}{date to, YYYY-MM-DD format}

\item{type_export}{Type of export like UPGCLOUD or TOGCLOUD.}

\item{email}{email of user}

\item{api_key}{Api Key obtain from /auth}

\item{dash_param}{AgroSpace internal use parameter. Default value FALSE}

\item{url}{URL for dev purpose}
}
\description{
This function set new task for downloading image indexes. This may take a some minutes.
}
\examples{
asapi_terramodels_post(client="clientexample", farm="farm1example",
sensor="S2SR", index="NDVI", date_start="2019-12-05", date_end="2021-02-07", type_export="UPGCLOUD", email="user.example@agrospace.cl", api_key=asapi_auth(email="user.example@agrospace.cl", password="contra1234")$api_key, url="https://api.agrospace.cl")

}
\keyword{API_KEY,}
\keyword{api_key}
