% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_tiles}
\alias{asapi_tiles}
\title{The image GET Function}
\usage{
asapi_tiles(
  client,
  farm,
  sensor,
  index,
  date,
  email,
  api_key,
  url = "https://api.agrospace.cl",
  dash_param = FALSE
)
}
\arguments{
\item{client}{Client name}

\item{farm}{Farm name to query}

\item{sensor}{Sensor}

\item{index}{Index}

\item{date}{imagen date}

\item{email}{email of user}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}

\item{dash_param}{AgroSpace internal use parameter. Default value FALSE}
}
\description{
This function brings a tiles to render directly in leaflet or QGIS
}
