% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_ts_image}
\alias{asapi_ts_image}
\title{The time series image GET Function}
\usage{
asapi_ts_image(
  client,
  farm,
  tableid,
  sensor,
  index,
  date_start,
  date_end,
  path = getwd(),
  email,
  api_key,
  url = "https://api.agrospace.cl",
  dash_param = FALSE
)
}
\arguments{
\item{client}{Client name.}

\item{farm}{Farm name to query.}

\item{tableid}{The mean values for the field (levelzero) or for each paddock (levelone)}

\item{sensor}{Sensor.}

\item{index}{Index.}

\item{date_start}{date from, YYYY-MM-DD format.}

\item{date_end}{date to, YYYY-MM-DD format.}

\item{path}{Address where the raster images will be saved. Default "getwd()".}

\item{email}{email of user.}

\item{api_key}{Api Key obtain from /auth.}

\item{url}{URL for dev purpose}

\item{dash_param}{AgroSpace internal use parameter. Default value FALSE.}
}
\description{
This function allows you to GET the time series (ts) of raster images.
}
