% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_url}
\alias{asapi_url}
\title{URL endpoint}
\usage{
asapi_url(url = "https://api.agrospace.cl", endpoint)
}
\arguments{
\item{url}{URL for dev purpose}

\item{endpoint}{Endpoint}
}
\description{
Generical function to translate json to data.frame
}
\examples{
asapi_url(url='https://api.agrospace.cl',endpoint='/farm')

}
