% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_user_get}
\alias{asapi_user_get}
\title{The Users GET Function}
\usage{
asapi_user_get(client, email, user, api_key, url = "https://api.agrospace.cl")
}
\arguments{
\item{client}{Client name}

\item{email}{email of user}

\item{user}{user name}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}
}
\description{
This function allows you to GET User information with the API and retrieve your API KEY
}
\examples{
asapi_user_get(client="clientexample", email="user.example@agrospace.cl", user="userexample", api_key="APIKEYEXAMPLE")
}
\keyword{API_KEY,}
\keyword{api_key}
