% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_user_post}
\alias{asapi_user_post}
\title{The Users POST Function}
\usage{
asapi_user_post(
  client,
  user_name,
  email_user,
  password,
  rol,
  email,
  api_key,
  url = "https://api.agrospace.cl"
)
}
\arguments{
\item{client}{Client name}

\item{user_name}{The new user name}

\item{email_user}{Email of the new user}

\item{password}{password of the new user}

\item{rol}{rol of the new user (Optional)}

\item{email}{email of user}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}
}
\description{
This function allows you to POST New User with the API and retrieve your API KEY
}
\examples{
asapi_user_post(client="clientexample", user_name="New User Example", email_user="newuser.example@agrospace.cl", password="Contra123", rol="user", email="user.example@agrospace.cl", api_key="APIKEYEXAMPLE")
}
\keyword{API_KEY,}
\keyword{api_key}
