% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_user_put}
\alias{asapi_user_put}
\title{The Users PUT Function}
\usage{
asapi_user_put(
  client,
  email,
  new_username = "NULL",
  new_email = "NULL",
  new_rol = "NULL",
  email_admin,
  api_key,
  url = "https://api.agrospace.cl"
)
}
\arguments{
\item{client}{Client name}

\item{email}{What is the email of the user to edit.}

\item{new_username}{New user name. If the value is 'NULL' it does not change.}

\item{new_email}{New email user. If the value is 'NULL' it does not change.}

\item{new_rol}{New user rol: admin or user. If the value is 'NULL' it does not change.}

\item{email_admin}{Email of the user who is editing. It must be 'admin'.}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}
}
\description{
This function allows you to PUT User new information with the API and retrieve your API KEY
}
\examples{
asapi_user_put(client="clientexample", email="user.example@agrospace.cl", new_username="EDIT New User", email_admin="user.example@agrospace.cl", api_key="APIKEYEXAMPLE")
}
\keyword{API_KEY,}
\keyword{api_key}
