% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tools.R
\name{get_rgb_plot}
\alias{get_rgb_plot}
\title{Plot RGB}
\usage{
get_rgb_plot(
  client,
  farm,
  date,
  email,
  api_key,
  url = "https://api.agrospace.cl",
  path = getwd(),
  name = NULL
)
}
\arguments{
\item{client}{Client name}

\item{farm}{email of user}

\item{date}{URL for dev purpose}

\item{email}{Email of customer user}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}

\item{path}{Address where the image will be saved. Default "getwd()"}

\item{name}{Name that the image will take. Default "client_farm_dare_RGB"}
}
\description{
This function allows you to check if the raster was download before in order to avoid repeated downloading of data
}
