% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tools.R
\name{read_shape}
\alias{read_shape}
\title{Read the Shape data from API}
\usage{
read_shape(client, farm, email, api_key, dash_param = FALSE, reset = FALSE)
}
\arguments{
\item{client}{Client name}

\item{farm}{email of user}

\item{email}{Email of custome user}

\item{api_key}{Api Key obtain from /auth}

\item{dash_param}{AgroSpace internal use parameter. Default value FALSE.}

\item{reset}{reset temporal folder}
}
\description{
This function allows you to check if the raster was downlaod before in order to avoid repeated downloading of data
}
