% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tools.R
\name{tmp_creator}
\alias{tmp_creator}
\title{Create temporal folder in main directory}
\usage{
tmp_creator(reset = FALSE)
}
\arguments{
\item{reset}{Reset all the tmp directory}
}
\description{
This function create a temporal folder in local directory
}
\examples{
tmp_creator()

}
