% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_counted}
\alias{asapi_counted}
\title{The Counted Calls GET Function}
\usage{
asapi_counted(client, email, api_key, url = "https://api.agrospace.cl")
}
\arguments{
\item{client}{Client name.}

\item{email}{Email of user.}

\item{api_key}{Api Key obtain from /auth.}

\item{url}{URL for dev purpose.}
}
\description{
This function get list of counted calls for a user.
Agrospace members only, with apikey master.
}
\examples{
asapi_counted(client="clientexample", email="user.example@agrospace.cl", api_key="APIKEYEXAMPLE")
}
\keyword{API_KEY,}
\keyword{api_key}
