% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_farm_post}
\alias{asapi_farm_post}
\title{The Farms POST Function}
\usage{
asapi_farm_post(
  client,
  farm_name,
  geojson = "NULL",
  email,
  api_key,
  url = "https://api.agrospace.cl"
)
}
\arguments{
\item{client}{Client name}

\item{email}{email of user}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}

\item{farm}{Farm name to query}
}
\description{
This function allows you to POST new Farm with the API and retrieve your API KEY
}
\examples{
asapi_farm_post(client="clientexample", farm_name="farm.example", email="user.example@agrospace.cl", api_key="APIKEYEXAMPLE")
}
\keyword{Farms}
