% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_image}
\alias{asapi_image}
\title{The image GET Function}
\usage{
asapi_image(
  client,
  farm,
  sensor,
  index,
  date,
  email,
  api_key,
  url = "https://api.agrospace.cl",
  dash_param = FALSE
)
}
\arguments{
\item{client}{Client name}

\item{farm}{Farm name to query}

\item{sensor}{Sensor}

\item{index}{Index}

\item{date}{imagen date}

\item{email}{email of user}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}

\item{dash_param}{AgroSpace internal use parameter. Default value FALSE}
}
\description{
This function allows you to GET Client - Farm raster information with the API
}
\examples{
asapi_image(client="clientexample", farm="farm1example",
sensor="S2SR", index="NDVI", date="2021-02-07", email="user.example@agrospace.cl", api_key=asapi_auth(email="user.example@agrospace.cl", password="contra1234")$api_key)

}
\keyword{API_KEY,}
\keyword{api_key}
