% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_layout_post}
\alias{asapi_layout_post}
\title{The Layout image POST Function}
\usage{
asapi_layout_post(
  client,
  path,
  farm,
  email,
  api_key,
  url = "https://api.agrospace.cl"
)
}
\arguments{
\item{client}{Client name.}

\item{path}{Address of the new logo.}

\item{farm}{Farm name.}

\item{email}{email of user}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}
}
\description{
This function upload a farms's layout with the API
}
