% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_logo_post}
\alias{asapi_logo_post}
\title{The Logo image POST Function}
\usage{
asapi_logo_post(client, path, email, api_key, url = "https://api.agrospace.cl")
}
\arguments{
\item{client}{Client name.}

\item{path}{Address of the new logo.}

\item{email}{email of user}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}
}
\description{
This function upload a client's logo with the API
}
