% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_table}
\alias{asapi_table}
\title{The Table GET Function}
\usage{
asapi_table(
  client,
  farm,
  tableid,
  sensor,
  index,
  date_start,
  date_end,
  email,
  api_key,
  dash_param = FALSE,
  url = "https://api.agrospace.cl"
)
}
\arguments{
\item{client}{Client name}

\item{farm}{Farm name to query}

\item{tableid}{The mean values for the field (levelzero) or for each paddock (levelone)}

\item{sensor}{Sensor}

\item{index}{Index}

\item{date_start}{date from, YYYY-MM-DD format}

\item{date_end}{date to, YYYY-MM-DD format}

\item{email}{email of user}

\item{api_key}{Api Key obtain from /auth}

\item{dash_param}{AgroSpace internal use parameter. Default value FALSE}

\item{url}{URL for dev purpose}
}
\description{
This function allows you to GET Client Farms  information with the API
}
\examples{
asapi_table(client="clientexample", farm="farm1example",
tableid="levelzero", sensor="S2SR", index="NDVI", date_start="2019-12-05", date_end="2021-02-07", email="user.example@agrospace.cl", api_key=asapi_auth(email="user.example@agrospace.cl", password="contra1234")$api_key, url="https://api.agrospace.cl")

}
\keyword{API_KEY,}
\keyword{api_key}
