% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_connector.R
\name{asapi_user_delete}
\alias{asapi_user_delete}
\title{The Users DELETE Function}
\usage{
asapi_user_delete(
  client,
  email_deleted,
  email,
  api_key,
  url = "https://api.agrospace.cl"
)
}
\arguments{
\item{client}{Client name}

\item{email_deleted}{User's email to delete}

\item{email}{email of user admin}

\item{api_key}{Api Key obtain from /auth}

\item{url}{URL for dev purpose}
}
\description{
This function allows you to DELETE User information with the API and retrieve your API KEY
}
\examples{
asapi_user_delete(client="clientexample", email_deleted="newuser.example@agrospace.cl", email="user.example@agrospace.cl", api_key=asapi_auth(email="user.example@agrospace.cl", password="contra1234")$api_key)
}
\keyword{API_KEY,}
\keyword{api_key}
