% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_tools.R
\name{read_rst}
\alias{read_rst}
\title{Read the Raster data from API}
\usage{
read_rst(
  client,
  farm,
  sensor,
  index,
  date,
  email,
  api_key,
  dash_param = FALSE,
  reset = FALSE
)
}
\arguments{
\item{client}{Client name}

\item{farm}{email of user}

\item{sensor}{satellite sensor}

\item{index}{Api Key obtain from /auth}

\item{date}{URL for dev purpose}

\item{email}{Email of custome user}

\item{api_key}{Api Key obtain from /auth}

\item{dash_param}{AgroSpace internal use parameter. Default value FALSE.}

\item{reset}{reset temporal folder}
}
\description{
This function allows you to check if the raster was downlaod before in order to avoid repeated downloading of data
}
\examples{
\dontrun{
read_rst(client='clientexample', farm='farm1example', sensor = "S2SR", index = "NDVI", date = '2021-02-07', email="user.example@agrospace.cl", api_key=asapi_auth(email="user.example@agrospace.cl", password="contra1234")$api_key)
}
}
\keyword{and}
\keyword{in}
\keyword{keep}
\keyword{local}
\keyword{raster}
\keyword{read}
